+++ cw-orch-daemon/src/core.rs
@@ -14,7 +14,7 @@ use cosmrs::{
 use cosmwasm_std::{Addr, Binary, Coin};
 use cw_orch_core::{
     contract::interface_traits::Uploadable,
-    environment::{ChainState, IndexResponse},
+    environment::{AsyncWasmQuerier, ChainState, IndexResponse, Querier},
     log::transaction_target,
 };
 use flate2::{write, Compression};
@@ -344,6 +344,32 @@ impl<Sender: SenderTrait> DaemonAsyncBase<Sender> {
         DaemonAsyncBase {
             sender,
             state: self.state,
+        }
+    }
+}
+
+impl Querier for DaemonAsync {
+    type Error = DaemonError;
+}
+
+impl AsyncWasmQuerier for DaemonAsync {
+    /// Query a contract.
+    fn smart_query<Q: Serialize + Sync, T: DeserializeOwned>(
+        &self,
+        address: impl Into<String> + Send,
+        query_msg: &Q,
+    ) -> impl std::future::Future<Output = Result<T, DaemonError>> + Send {
+        let query_data = serde_json::to_vec(&query_msg).unwrap();
+        async {
+            let mut client =
+                cosmos_modules::cosmwasm::query_client::QueryClient::new(self.channel());
+            let resp = client
+                .smart_contract_state(cosmos_modules::cosmwasm::QuerySmartContractStateRequest {
+                    address: address.into(),
+                    query_data,
+                })
+                .await?;
+            Ok(from_str(from_utf8(&resp.into_inner().data).unwrap())?)
         }
     }
 }